requirements.txt
ExifRead==2.3.2
Pillow==8.2.0
